move --env-keep option to avocado_runner_remote
The env-keep plugin was only creating an option (--env-keep) that was
only used by one plugin (runner_remote). This is not migrating env-keep
to the new model, only removing env-keep plugin and migrating this
option to the optional_plugin/runner_remote.
